+++ lib/redmine_expert_helpdesk/api_serializers.rb
@@ -247,6 +247,7 @@ module RedmineExpertHelpdesk
         api.info_request_require_attachment s.info_request_require_attachment
         api.info_request_min_attachment_kb  s.info_request_min_attachment_kb
         api.info_request_keywords           s.info_request_keywords
+        api.info_request_sender_blacklist   s.info_request_sender_blacklist
         api.info_request_threshold          s.info_request_threshold
         api.info_request_ai_prompt_mode     s.info_request_ai_prompt_mode
         api.info_request_ai_prompt          s.info_request_ai_prompt